V0.0.4-2022-04-22
基类调整
1. filters提取过滤常量变量
2. 提供Where类以支持统一的自定义过滤方式
3. 修复当通过请求注入过滤条件时，同一字段多种过滤不生效的问题
4. 请求注入过滤支持NotIn过滤

dbengine调整
1. UnionServiceMixin逻辑优化

parser调整
1. parser全局异常捕获

+++ src/lesoon_restful/dbengine/alchemy/wrappers/service.py
@@ -25,15 +25,21 @@ class UnionServiceMixin:
                       delete_rows: t.List[int],
                       commit: bool = True):
         """新增，更新，删除的联合操作."""
-        self.create_many(items=self.schema.load(insert_rows, many=True),
-                         commit=False)
-        self.update_many(items=[
-            self.read_or_raise(r.get(self.id_attribute)) for r in update_rows
-        ],
-                         changes=update_rows,
-                         commit=False)
-        self.delete_many(ids=delete_rows, commit=False)
-        self.commit_or_flush(commit)
+        if insert_rows:
+            self.create_many(items=self.schema.load(insert_rows, many=True),
+                             commit=False)
+        if update_rows:
+            self.update_many(items=[
+                self.read_or_raise(r.get(self.id_attribute))
+                for r in update_rows
+            ],
+                             changes=update_rows,
+                             commit=False)
+        if delete_rows:
+            self.delete_many(ids=delete_rows, commit=False)
+
+        if insert_rows or update_rows or delete_rows:
+            self.commit_or_flush(commit)
 
 
 class ComplexServiceMixin:
